Разложил по полочкам
Все классы необходимые для игры находятся в своих файлах. Короче, иерархия
