+++ settings.py
@@ -1,0 +1,15 @@
+import pygame
+
+WIDTH = 800
+HEIGHT = 600
+FPS = 45
+
+# Задаем цвета
+WHITE = (255, 255, 255)
+BLACK = (0, 0, 0)
+RED = (255, 0, 0)
+GREEN = (0, 255, 0)
+BLUE = (0, 0, 255)
+YELLOW = (255, 255, 0)
+
+all_sprites = pygame.sprite.Group()
